+++ modules/tinymce/src/core/main/ts/delete/BlockRangeDelete.ts
@@ -5,7 +5,7 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Fun, Optionals } from '@ephox/katamari';
+import { Fun, Optional, Optionals } from '@ephox/katamari';
 import { Compare, PredicateFind, SugarElement } from '@ephox/sugar';
 
 import EditorSelection from '../api/dom/Selection';
@@ -16,7 +16,7 @@ import * as ElementType from '../dom/ElementType';
 import * as DeleteUtils from './DeleteUtils';
 import * as MergeBlocks from './MergeBlocks';
 
-const deleteRangeMergeBlocks = (rootNode: SugarElement<Node>, selection: EditorSelection): boolean => {
+const deleteRangeMergeBlocks = (rootNode: SugarElement<Node>, selection: EditorSelection): Optional<() => void> => {
   const rng = selection.getRng();
 
   return Optionals.lift2(
@@ -24,17 +24,18 @@ const deleteRangeMergeBlocks = (rootNode: SugarElement<Node>, selection: EditorS
     DeleteUtils.getParentBlock(rootNode, SugarElement.fromDom(rng.endContainer)),
     (block1, block2) => {
       if (Compare.eq(block1, block2) === false) {
-        rng.deleteContents();
+        return Optional.some(() => {
+          rng.deleteContents();
 
-        MergeBlocks.mergeBlocks(rootNode, true, block1, block2).each((pos) => {
-          selection.setRng(pos.toRange());
+          MergeBlocks.mergeBlocks(rootNode, true, block1, block2).each((pos) => {
+            selection.setRng(pos.toRange());
+          });
+
         });
-
-        return true;
       } else {
-        return false;
+        return Optional.none();
       }
-    }).getOr(false);
+    }).getOr(Optional.none());
 };
 
 const isRawNodeInTable = (root: SugarElement<Node>, rawNode: Node): boolean => {
@@ -52,20 +53,21 @@ const isEverythingSelected = (root: SugarElement<Node>, rng: Range): boolean => 
   return !isSelectionInTable(root, rng) && noPrevious && noNext;
 };
 
-const emptyEditor = (editor: Editor): boolean => {
-  editor.setContent('');
-  editor.selection.setCursorLocation();
-  return true;
+const emptyEditor = (editor: Editor): Optional<() => void> => {
+  return Optional.some(() => {
+    editor.setContent('');
+    editor.selection.setCursorLocation();
+  });
 };
 
-const deleteRange = (editor: Editor): boolean => {
+const deleteRange = (editor: Editor): Optional<() => void> => {
   const rootNode = SugarElement.fromDom(editor.getBody());
   const rng = editor.selection.getRng();
   return isEverythingSelected(rootNode, rng) ? emptyEditor(editor) : deleteRangeMergeBlocks(rootNode, editor.selection);
 };
 
-const backspaceDelete = (editor: Editor, _forward: boolean): boolean =>
-  editor.selection.isCollapsed() ? false : deleteRange(editor);
+const backspaceDelete = (editor: Editor, _forward: boolean): Optional<() => void> =>
+  editor.selection.isCollapsed() ? Optional.none() : deleteRange(editor);
 
 export {
   backspaceDelete