+++ modules/tinymce/src/core/main/ts/keyboard/MatchKeys.ts
@@ -7,26 +7,44 @@
 
 import { Arr, Fun, Optional } from '@ephox/katamari';
 
-export interface KeyPattern {
+interface KeyPatternBase {
   shiftKey?: boolean;
   altKey?: boolean;
   ctrlKey?: boolean;
   metaKey?: boolean;
   keyCode?: number;
-  action: () => boolean;
 }
 
-const defaultPatterns = (patterns: KeyPattern[]): KeyPattern[] => Arr.map(patterns, (pattern) => ({
+export interface KeyPattern extends KeyPatternBase {
+  action: () => boolean;
+}
+export interface KeyPatternDelayed extends KeyPatternBase {
+  action: () => Optional<() => void>;
+}
+
+const baseKeyPattern = {
   shiftKey: false,
   altKey: false,
   ctrlKey: false,
   metaKey: false,
-  keyCode: 0,
-  action: Fun.noop,
-  ...pattern
-}));
+  keyCode: 0
+};
 
-const matchesEvent = (pattern: KeyPattern, evt: KeyboardEvent) => (
+const defaultPatterns = (patterns: KeyPattern[]): KeyPattern[] =>
+  Arr.map(patterns, (pattern) => ({
+    ...baseKeyPattern,
+    action: Fun.noop,
+    ...pattern
+  }));
+
+const defaultDelayedPatterns = (patterns: KeyPatternDelayed[]): KeyPatternDelayed[] =>
+  Arr.map(patterns, (pattern) => ({
+    ...baseKeyPattern,
+    action: () => Optional.none(),
+    ...pattern
+  }));
+
+const matchesEvent = <T extends KeyPatternBase>(pattern: T, evt: KeyboardEvent) => (
   evt.keyCode === pattern.keyCode &&
   evt.shiftKey === pattern.shiftKey &&
   evt.altKey === pattern.altKey &&
@@ -37,13 +55,20 @@ const matchesEvent = (pattern: KeyPattern, evt: KeyboardEvent) => (
 const match = (patterns: KeyPattern[], evt: KeyboardEvent) =>
   Arr.bind(defaultPatterns(patterns), (pattern) => matchesEvent(pattern, evt) ? [ pattern ] : [ ]);
 
+const matchDelayed = (patterns: KeyPatternDelayed[], evt: KeyboardEvent) =>
+  Arr.bind(defaultDelayedPatterns(patterns), (pattern) => matchesEvent(pattern, evt) ? [ pattern ] : [ ]);
+
 const action = <T extends (...args: any[]) => any>(f: T, ...x: Parameters<T>) => (): ReturnType<T> => f.apply(null, x);
 
 const execute = (patterns: KeyPattern[], evt: KeyboardEvent): Optional<KeyPattern> =>
   Arr.find(match(patterns, evt), (pattern) => pattern.action());
 
+const executeWithDelayedAction = (patterns: KeyPatternDelayed[], evt: KeyboardEvent): Optional<() => void> =>
+  Arr.findMap(matchDelayed(patterns, evt), (pattern) => pattern.action());
+
 export {
   match,
   action,
-  execute
+  execute,
+  executeWithDelayedAction
 };