+++ modules/tinymce/src/themes/silver/main/ts/backstage/Backstage.ts
@@ -56,7 +56,7 @@ export interface UiFactoryBackstage {
   setContextMenuState?: (state: boolean) => void;
 }
 
-const init = (sink: AlloyComponent, editor: Editor, lazyAnchorbar: () => AlloyComponent): UiFactoryBackstage => {
+const init = (lazySink: () => Result<AlloyComponent, string>, editor: Editor, lazyAnchorbar: () => AlloyComponent): UiFactoryBackstage => {
   const contextMenuState = Cell(false);
   const toolbar = HeaderBackstage(editor);
   const backstage: UiFactoryBackstage = {
@@ -65,13 +65,13 @@ const init = (sink: AlloyComponent, editor: Editor, lazyAnchorbar: () => AlloyCo
         icons: () => editor.ui.registry.getAll().icons,
         menuItems: () => editor.ui.registry.getAll().menuItems,
         translate: I18n.translate,
-        isDisabled: () => editor.mode.isReadOnly() || editor.ui.isDisabled(),
+        isDisabled: () => editor.mode.isReadOnly() || !editor.ui.isEnabled(),
         getOption: editor.options.get
       },
       interpreter: (s) => UiFactory.interpretWithoutForm(s, {}, backstage),
       anchors: Anchors.getAnchors(editor, lazyAnchorbar, toolbar.isPositionedAtTop),
       header: toolbar,
-      getSink: () => Result.value(sink)
+      getSink: lazySink
     },
     urlinput: UrlInputBackstage(editor),
     styles: initStyleFormatBackstage(editor),